refactor: GachaTicketConsumeAmount移動し忘れ
